blog: Render blog entries as Jinja2 templates
This makes it possible to use e.g. template functions and filters in
the blog entries.

For instance, this commit adds a function entry_url() that, given a
slug of an existing blog entry, returns the absolute url to that
entry.

While at it, simplify the blog view code a bit.